% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svar_functions.R
\name{data.transform}
\alias{data.transform}
\title{data.transform}
\usage{
data.transform(nos, nov, data.list, fwd, variable.name, shock.name)
}
\arguments{
\item{nos}{number of shocks}

\item{nov}{number of dependent variables}

\item{data.list}{a list containing all the IRF draws}

\item{fwd}{forward steps in impulse response function}

\item{variable.name}{a character vector containing name of dependent variables}

\item{shock.name}{a character vector containing shock names}
}
\value{
the function returns a dataframe containing the impulse response function with user-defined forward steps
}
\description{
transform the list containing all the impulse response functions
into a dataframe with flat structure
}

