% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svar_functions.R
\name{find.base.draw}
\alias{find.base.draw}
\title{find.base.draw}
\usage{
find.base.draw(irf.all)
}
\arguments{
\item{irf.all}{A list containing all the IRF draws}
}
\value{
the function returns the baseline draw index
}
\description{
find the baseline draw whose corresponding impulse responses
has the minimimum sum of distance squared to the median value
}

