% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svar_functions.R
\name{generateDraw}
\alias{generateDraw}
\title{generateDraw}
\usage{
generateDraw(C, k)
}
\arguments{
\item{C}{initial short run impact matrix}

\item{k}{number of dependent variables}
}
\value{
the function returns a randomised initial short-run impact matrix
}
\description{
generate a orthogonal matrix by taking the QR
decomposition of a random matrix, and post multiply
the initial matrix by this random orthogonal draw
}

