% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svar_functions.R
\name{hist.decomp}
\alias{hist.decomp}
\title{hist.decomp}
\usage{
\method{hist}{decomp}(Z, u, B, X.mat, nov, nos, t.last, var.name, shock.name)
}
\arguments{
\item{Z}{the matrix that transforms structral shocks to shocks}

\item{u}{model residuals}

\item{B}{estimated coefficient matrix}

\item{X.mat}{data for the predictors}

\item{nov}{number of variable}

\item{nos}{number of shocks}

\item{var.name}{variable names}

\item{shock.name}{shock names}
}
\value{
the function returns an array containing the historical decompositions 
        for the model residuals.
        
        The dimensions of the array are response variable, shocks, time respectively
}
\description{
find the historical decomposition for model residuals.
}

