% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svar_functions.R
\name{usvar}
\alias{usvar}
\title{usvar}
\usage{
usvar(p, k, f, sr, B, C1, t, draws, var_pos)
}
\arguments{
\item{p}{number of lags in the model}

\item{k}{number of dependent Variables}

\item{f}{zero restriction matrix}

\item{sr}{sign restriction matrix}

\item{B}{coefficient matrix}

\item{C1}{initial impact response}

\item{t}{length of impulse response function}

\item{draws}{Number of draws to find}

\item{var_pos}{...}
}
\value{
the function returns a list that stores the impulse response for every draw,
        and each draw is a nested list that contains the impulse response matrix
        for each response variable.
}
\description{
Solve a underidentified Structural Autogressive models SVAR models
}

