% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svar_functions.R
\name{xlags}
\alias{xlags}
\title{xlags}
\usage{
xlags(data, p)
}
\arguments{
\item{data}{A matrix containing data for lagging, where each column contains time series data for a variable}

\item{p}{Number of lags}
}
\value{
the function returns a data matrix after lag operation
}
\description{
Create lagged values for a data matrix where every column contains data for a variable
}

